RISC_OSLib: txtar: Support extended scroll requests
Use extended scroll requests to handle mouse wheel scrolling, to avoid
the "thumbing" code getting confused and thinking the user is dragging
the scrollbar (the "thumbing" mode of scrolling isn't desirable for the
mouse wheel, because it largely ignores the requested scroll distance)

This change allows Edit & SrcEdit (and anything else using txtar) to
handle the scroll wheel better, now that the WindowScroll module is
controlling things.

Version 6.06. Tagged as 'RISC_OSLib-6_06'